+++ Tunny/Handler/ProgressState.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace Tunny.Handler
@@ -9,5 +10,6 @@ namespace Tunny.Handler
         public int ObjectiveNum { get; set; }
         public double[][] BestValues { get; set; }
         public double HypervolumeRatio { get; set; }
+        public TimeSpan EstimatedTimeRemaining { get; set; }
     }
 }